create edit button in the reviews component
